Fix some API connection problems
