+++ mentor_booking/.gitignore
@@ -31,3 +31,4 @@ build/
 
 ### VS Code ###
 .vscode/
+/.apt_generated_tests/
