+++ mentor_booking/src/main/java/com/project/MentorBookingApplication.java
@@ -1,0 +1,14 @@
+package com.project;
+
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class MentorBookingApplication {
+    
+    public static void main(String[] args) {
+        SpringApplication.run(MentorBookingApplication.class, args);
+    }
+
+}
