+++ mentor_booking/src/main/resources/application.properties
@@ -5,8 +5,9 @@ spring.datasource.username=root
 spring.datasource.password=
 
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.SQLServerDialect
-spring.jpa.hibernate.ddl-auto=create
+spring.jpa.hibernate.ddl-auto=none
 spring.jpa.hibernate.naming.physical-strategy=org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
 
 
 spring.session.jdbc.initialize-schema: always
+
